+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed:
 - Properly checks for existing blinded proposals
 - Uses pokedex's verifier registry
+- Make `blockNumber`, `counts` and `roots` state variables publicly readable
 
 ## [0.1.0] - 2018-04-12
 ### Added: